requirements.txt
langgraph==0.0.64
langchain==0.1.9
langchain-openai==0.0.5
openai==1.3.9
streamlit==1.28.1
python-dotenv==1.0.0
httpx==0.25.1
requests==2.31.0
newsapi==0.1.3
